تعديل منطق التطبيق لتكون لكل شركة صلاحيات وتعديل منطق ال payrole والغء double في الماليات وانشاء نظام hr wydv

+++ StoreManagement/StoreManagement.Infrastructure/Services/CashTransactionService.cs
@@ -2,7 +2,14 @@ using Microsoft.EntityFrameworkCore;
 using StoreManagement.Data;
 using StoreManagement.Shared.Common;
 using StoreManagement.Shared.DTOs;
-using StoreManagement.Shared.Entities;
+using StoreManagement.Shared.Entities.HR;
+using StoreManagement.Shared.Entities.Inventory;
+using StoreManagement.Shared.Entities.Sales;
+using StoreManagement.Shared.Entities.Finance;
+using StoreManagement.Shared.Entities.Identity;
+using StoreManagement.Shared.Entities.Partners;
+using StoreManagement.Shared.Entities.Configuration;
+using StoreManagement.Shared.Entities.Core;
 using StoreManagement.Shared.Enums;
 using StoreManagement.Shared.Interfaces;
 
@@ -190,8 +197,8 @@ public class CashTransactionService : ICashTransactionService
     {
         if (!transaction.RelatedEntityId.HasValue) return;
 
-        double factor = reverse ? -1 : 1;
-        double amount = transaction.Value * factor;
+        decimal factor = reverse ? -1 : 1;
+        decimal amount = transaction.Value * factor;
 
         if (transaction.SourceType == TransactionSource.Customer)
         {
